consolidate registry source files
