+++ internal/experimental/registry/client_test.go
