--- internal/experimental/registry/registry_op_pull_options.go
@@ -1,49 +1,0 @@
-/*
-Copyright The Helm Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package registry // import "helm.sh/helm/v3/internal/experimental/registry"
-
-type (
-	// PullOption allows specifying various settings on pull
-	PullOption func(*pullOperation)
-
-	pullOperation struct {
-		withChart         bool
-		withProv          bool
-		ignoreMissingProv bool
-	}
-)
-
-// PullOptWithChart returns a function that sets the withChart setting on pull
-func PullOptWithChart(withChart bool) PullOption {
-	return func(operation *pullOperation) {
-		operation.withChart = withChart
-	}
-}
-
-// PullOptWithProv returns a function that sets the withProv setting on pull
-func PullOptWithProv(withProv bool) PullOption {
-	return func(operation *pullOperation) {
-		operation.withProv = withProv
-	}
-}
-
-// PullOptIgnoreMissingProv returns a function that sets the ignoreMissingProv setting on pull
-func PullOptIgnoreMissingProv(ignoreMissingProv bool) PullOption {
-	return func(operation *pullOperation) {
-		operation.ignoreMissingProv = ignoreMissingProv
-	}
-}
