--- internal/experimental/registry/registry_op_push.go
@@ -1,93 +1,0 @@
-/*
-Copyright The Helm Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package registry // import "helm.sh/helm/v3/internal/experimental/registry"
-
-import (
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
-
-	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
-	"oras.land/oras-go/pkg/content"
-	"oras.land/oras-go/pkg/oras"
-)
-
-// Push uploads a chart to a registry.
-func (c *Client) Push(data []byte, ref string, options ...PushOption) (*PushResult, error) {
-	operation := &pushOperation{
-		strictMode: true, // By default, enable strict mode
-	}
-	for _, option := range options {
-		option(operation)
-	}
-	meta, err := extractChartMeta(data)
-	if err != nil {
-		return nil, err
-	}
-	if operation.strictMode {
-		if !strings.HasSuffix(ref, fmt.Sprintf("/%s:%s", meta.Name, meta.Version)) {
-			return nil, errors.New(
-				"strict mode enabled, ref basename and tag must match the chart name and version")
-		}
-	}
-	store := content.NewMemoryStore()
-	chartDescriptor := store.Add("", ChartLayerMediaType, data)
-	configData, err := json.Marshal(meta)
-	if err != nil {
-		return nil, err
-	}
-	configDescriptor := store.Add("", ConfigMediaType, configData)
-	descriptors := []ocispec.Descriptor{chartDescriptor}
-	var provDescriptor ocispec.Descriptor
-	if operation.provData != nil {
-		provDescriptor = store.Add("", ProvLayerMediaType, operation.provData)
-		descriptors = append(descriptors, provDescriptor)
-	}
-	manifest, err := oras.Push(ctx(c.out, c.debug), c.resolver, ref, store, descriptors,
-		oras.WithConfig(configDescriptor), oras.WithNameValidation(nil))
-	if err != nil {
-		return nil, err
-	}
-	chartSummary := &descriptorPushSummaryWithMeta{
-		Meta: meta,
-	}
-	chartSummary.Digest = chartDescriptor.Digest.String()
-	chartSummary.Size = chartDescriptor.Size
-	result := &PushResult{
-		Manifest: &descriptorPushSummary{
-			Digest: manifest.Digest.String(),
-			Size:   manifest.Size,
-		},
-		Config: &descriptorPushSummary{
-			Digest: configDescriptor.Digest.String(),
-			Size:   configDescriptor.Size,
-		},
-		Chart: chartSummary,
-		Prov:  &descriptorPushSummary{}, // prevent nil references
-		Ref:   ref,
-	}
-	if operation.provData != nil {
-		result.Prov = &descriptorPushSummary{
-			Digest: provDescriptor.Digest.String(),
-			Size:   provDescriptor.Size,
-		}
-	}
-	fmt.Fprintf(c.out, "Pushed: %s\n", result.Ref)
-	fmt.Fprintf(c.out, "Digest: %s\n", result.Manifest.Digest)
-	return result, err
-}
