--- internal/experimental/registry/registry_op_push_options.go
@@ -1,41 +1,0 @@
-/*
-Copyright The Helm Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package registry // import "helm.sh/helm/v3/internal/experimental/registry"
-
-type (
-	// PushOption allows specifying various settings on push
-	PushOption func(*pushOperation)
-
-	pushOperation struct {
-		provData   []byte
-		strictMode bool
-	}
-)
-
-// PushOptProvData returns a function that sets the prov bytes setting on push
-func PushOptProvData(provData []byte) PushOption {
-	return func(operation *pushOperation) {
-		operation.provData = provData
-	}
-}
-
-// PushOptStrictMode returns a function that sets the strictMode setting on push
-func PushOptStrictMode(strictMode bool) PushOption {
-	return func(operation *pushOperation) {
-		operation.strictMode = strictMode
-	}
-}
