--- internal/experimental/registry/registry_op_push_result.go
@@ -1,40 +1,0 @@
-/*
-Copyright The Helm Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package registry // import "helm.sh/helm/v3/internal/experimental/registry"
-
-import "helm.sh/helm/v3/pkg/chart"
-
-type (
-	// PushResult is the result returned upon successful push.
-	PushResult struct {
-		Manifest *descriptorPushSummary         `json:"manifest"`
-		Config   *descriptorPushSummary         `json:"config"`
-		Chart    *descriptorPushSummaryWithMeta `json:"chart"`
-		Prov     *descriptorPushSummary         `json:"prov"`
-		Ref      string                         `json:"ref"`
-	}
-
-	descriptorPushSummary struct {
-		Digest string `json:"digest"`
-		Size   int64  `json:"size"`
-	}
-
-	descriptorPushSummaryWithMeta struct {
-		descriptorPushSummary
-		Meta *chart.Metadata `json:"meta"`
-	}
-)
